+++ src/map/mob.hpp
@@ -194,7 +194,7 @@ struct s_mob_skill {
 	int casttime,delay;
 	short cancel;
 	short cond1;
-	int64 cond2;
+	int cond2;
 	short target;
 	int val[5];
 	short emotion;
@@ -503,7 +503,7 @@ int mob_warpslave(struct block_list *bl, int range);
 int mob_linksearch(struct block_list *bl,va_list ap);
 
 bool mob_chat_display_message (mob_data &md, uint16 msg_id);
-int mobskill_use(struct mob_data *md,t_tick tick,int event, int64 damage = 0);
+int mobskill_use(struct mob_data *md,t_tick tick,int event, int damage = 0);
 int mobskill_event(struct mob_data *md,struct block_list *src,t_tick tick, int flag);
 int mob_summonslave(struct mob_data *md2,int *value,int amount,uint16 skill_id);
 int mob_countslave(struct block_list *bl);
